[refactoring]: all dсs objects moved to acceptance/dсs package
